Remove old files/directories and move some components
