+++ src/components/character/CharacterCard/styled.tsx
