+++ src/components/design-system/Button/types.ts
@@ -1,0 +1,10 @@
+import { ReactNode } from 'react'
+
+export type ButtonSize = 'small' | 'regular' | 'large'
+
+export type ButtonProps = {
+  children: ReactNode
+  disabled?: boolean
+  onClick: () => void
+  size?: ButtonSize
+}
