+++ src/components/design-system/Pagination/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
-import Button from '../../elements/Button'
-import FlexBox from '../../design-system/FlexBox'
-import Text from '../../design-system/Text'
-import Spacer from '../../design-system/Spacer'
-import { PaginationStyled } from './styled'
+import Button from '../Button'
+import FlexBox from '../FlexBox'
+import Text from '../Text'
+import Spacer from '../Spacer'
+import {PaginationStyled} from './styled'
 
 export interface PaginationBarProps {
   currentPage: number
@@ -15,20 +15,20 @@ export interface PaginationBarProps {
 }
 
 const Pagination: React.FC<PaginationBarProps> = ({
-  currentPage,
-  isNextDisabled,
-  isPreviousDisable,
-  lastPage,
-  onClickNext,
-  onClickPrevious,
-}) => {
+                                                    currentPage,
+                                                    isNextDisabled,
+                                                    isPreviousDisable,
+                                                    lastPage,
+                                                    onClickNext,
+                                                    onClickPrevious,
+                                                  }) => {
   return (
     <PaginationStyled>
       <FlexBox alignItems="center" justifyContent="center" fullWidth>
         <Button onClick={onClickPrevious} disabled={isPreviousDisable}>
           <Text typography="normal">Previous</Text>
         </Button>
-        <Spacer level={3} isHorizontal />
+        <Spacer level={3} isHorizontal/>
         <Text typography="normal" usage="bold">
           {currentPage}
         </Text>
@@ -37,7 +37,7 @@ const Pagination: React.FC<PaginationBarProps> = ({
             &nbsp;{`/ ${lastPage}`}
           </Text>
         )}
-        <Spacer level={3} isHorizontal />
+        <Spacer level={3} isHorizontal/>
         <Button onClick={onClickNext} disabled={isNextDisabled}>
           <Text typography="normal">Next</Text>
         </Button>
