--- src/components/elements/Button/types.ts
@@ -1,1 +1,0 @@
-export type ButtonSize = 'small' | 'regular' | 'large'
