--- src/types/general/index.ts
@@ -1,3 +1,0 @@
-import { MouseEvent } from 'react'
-
-export type EventOnClickButton = MouseEvent<HTMLButtonElement, MouseEvent>
