--- src/types/index.ts
@@ -1,3 +1,0 @@
-export type { CharacterStatusType, CharacterGender } from './characters'
-
-export * from './general'
